feat: 댓글 수정 기능 구현

+++ backend/src/main/java/com/digginroom/digginroom/service/CommentService.java
@@ -7,7 +7,7 @@ import com.digginroom.digginroom.controller.dto.CommentResponse;
 import com.digginroom.digginroom.controller.dto.CommentsResponse;
 import com.digginroom.digginroom.domain.Comment;
 import com.digginroom.digginroom.domain.Member;
-import com.digginroom.digginroom.domain.Room;
+import com.digginroom.digginroom.exception.CommentException.NotSameRoomException;
 import com.digginroom.digginroom.repository.CommentRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -28,24 +28,40 @@ public class CommentService {
                 .toList());
     }
 
-    public CommentResponse comment(final Room room, final Member member, final CommentRequest request) {
-        Comment comment = new Comment(room.getId(), request.comment(), member);
+    public CommentResponse comment(final Long roomId, final Member member, final CommentRequest request) {
+        Comment comment = new Comment(roomId, request.comment(), member);
         commentRepository.save(comment);
 
         return CommentResponse.of(comment);
     }
 
-    public void delete(final Room room, final Member member, final Long commentId) {
-        Comment comment = commentRepository.findCommentByRoomIdAndId(room.getId(), commentId);
+    public void delete(final Long roomId, final Member member, final Long commentId) {
+        Comment comment = commentRepository.getCommentById(commentId);
 
-        isOwner(comment, member);
-
+        validateSameRoomAndOwner(member, roomId, comment);
         commentRepository.delete(comment);
     }
 
-    private void isOwner(final Comment comment, final Member member) {
+    public Comment update(final Member member, final Long roomId, final Long commentId, final CommentRequest request) {
+        Comment comment = commentRepository.getCommentById(commentId);
+        validateSameRoomAndOwner(member, roomId, comment);
+        return comment.updateComment(request.comment());
+    }
+
+    private void validateSameRoomAndOwner(final Member member, final Long roomId, final Comment comment) {
+        validateSameRoom(roomId, comment);
+        validateSameOwner(member, comment);
+    }
+
+    private void validateSameOwner(final Member member, final Comment comment) {
         if (!comment.isOwner(member)) {
             throw new NotOwnerException();
+        }
+    }
+
+    private void validateSameRoom(final Long roomId, final Comment comment) {
+        if (!comment.isSameRoom(roomId)) {
+            throw new NotSameRoomException();
         }
     }
 }
